% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuRbonegro.R
\name{tuRbonegro}
\alias{tuRbonegro}
\title{tuRbonegro
This fonction plays a randomly selected Turbonegro clip.}
\usage{
tuRbonegro(width = 560, height = 315, autoplay = TRUE)
}
\arguments{
\item{width}{width of the video, in pixel. Default is 560.}

\item{height}{width of the video, in pixel. Default is 315.}

\item{autoplay}{set video autoplay. Default is TRUE.}
}
\value{
plays a clip in you RStudio Viewer
}
\description{
tuRbonegro
This fonction plays a randomly selected Turbonegro clip.
}
\examples{
tuRbonegro()

}
