% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuRbonegro.R
\name{tuRbonegro}
\alias{tuRbonegro}
\title{tuRbonegro
This fonction plays a randomly selected Turbonegro clip.}
\usage{
tuRbonegro()
}
\value{
plays a clip in you RStudio Viewer
}
\description{
tuRbonegro
This fonction plays a randomly selected Turbonegro clip.
}
\examples{
tuRbonegro()

}
